hw1/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.3
